Removal of dedicated gameState component
Upon collision of a bird with obstacles, we just remove `bird` component
from entity representing a bird. Absence of entitites with `bird`
component acts as a condition for game over behaviour for game systems.

To preserve bouncing behaviour when a "dead" bird hits the ground,
collision system now look for entities with `body` components so it
continues to function when game is over, since bird entity does not
lose `body` component after collision.